Used a global buffer to manage string results (#11)
This is perparatory work necessary for #11
To_String functions for W_Object are now Write_String, writing strings to a global
buffer and returning the slice of that buffer used.
This will allow to track line / column of the buffer currently being created.
This also optimizes a lot concatenation happening in these calls.