Making progress with the indentation engine (#11)
buffer_column is now updated as the indentation is computed.
This requires the indentation to be a bit less user-friendly, and in particular to
identify potential start of lines to be idented with i"". These will actually be indented
only when confirmed to be start of lines (e.g. happening after an end of line character
in the buffer).