Simplified implementation and semantics of node kind predicates.
Writing a node kind predicate of e.g. the form Entity (<some expression>) used
to push an instrinsic call if "Entity" matched the kind of the langkit node,
said call checking the expression. The confusing consequence is that, even
if something like "match Entity" could match, it was stack a call and not
the value of the object, while "match Entity ()" would execute the call
and return the value. It turns out that entities already provide ways
to check their value against an expression, so stacking the object instead
of an instrinsic call already works for both cases.