Further cleanup and documentation for expressions.
One of these cleanup turned out to be the behavior of allocators in the
context of calls that are generating all values, e.g. in:
   child (new (something ())).all ()
Turns out there's no reason to forbid this, the natural semantic is that a
new object will be created for all child of the current it.