Implemented wider support for Ada properties and container iteration.
This patch extracts the property-generic processing of input kit into

separate packages. Only support is provided for Ada. A number of parameter
and returned types kinds are now supported (a few are missing and could need
to be completed in the future).

Actual parameter lists have been modified to accomodate a format {index,
value} used in this new processing.

Implements T930-056.