Fixed fold expression evaluation and error message.
The fold expression evaluated missed pushing the new implicit it, which
resulted sometimes in unexpected error down the line.

With this change, we're now getting proper error when e.g. the selector
of a property is wrong. That error message has been fixed to provide more
information of the type found.

Part of TA05-016